cuda: 修复 MMQ 目标偏移 int32 溢出（大词表 × 大批量）
根因：MMQ host 侧 `offset_dst` 为 32-bit 有符号整型，lm_head（tied
`token_embd.weight[4096,151669]`）在 `n ≥ 14209` 时溢出。`stride_col_dst = vocab
= 151669`、`J = 128`，当 `jt ≥ 111`（即 `ntx = ceil(n/128) ≥ 112`）时
`jt*J*stride_col_dst = 111*128*151669 = 2,154,913,152 > 2^31-1` 溢出为负 →
写回 `dst − ~8.6 GiB` 野地址 → 未映射页 Memory Fault
（`kernel: mul_mat_q<(ggml_type)103,128,true>`）。

修复：`ggml/src/ggml-cuda/mmq.cuh` 四处同型位置全部把声明改为 `int64_t` 并在乘法
显式 cast `(int64_t) jt*J*stride_col_dst`（只改声明不够，RHS 仍按 int 计算）：
- :1094 常规 dense MMQ
- :1182 stream-k 循环内（ids_dst/MoE 分支）
- :1271 stream-k 尾段入口
- :1414 stream-k fixup 内核（`+ it*I` 同一表达式内一并 64 位）

`offset_dst = 0;`（:1108/:1196/:1285）与 `offset_dst += it*I;`（:1132/:1226/:1310）
随类型自然适配，逐处核对无误；未改动 `offset_y`/`offset_x`/`write_back` 等其它量。

验证（A 机 gfx1151，HEAD d3e63dbfc）：
- 修前最小复现（repro/libs 旧库快照）：`m=151669, k=4096` 时 `n=14208` 安全、
  `n=14209` 必崩（Memory Fault，exit 134），阈值精确到 1；关阀
  `GGML_HIP_NO_ROCMFPX_MMQ=1` 不崩。
- 修后最小复现全矩阵（n ∈ {512,4096,14208,14209,15232,16384,20000,24576,32768}，
  k ∈ {1024,4096}，含 m=251/4096 对照）全部不崩；数值对照（MMQ vs 关阀 hipBLAS）
  在 n ∈ {512,14208,14209,15232,16384,20000} 逐元素一致
  （max abs diff 3.557e2、NMSE 5.46e-5，与 n=512 安全基线同量级）。n=32768 的
  hipBLAS 对照臂因 rocBLAS `invalid configuration argument` 失败（既有限制，修前
  同样存在，非本次引入），MMQ 修复路径自身通过。
- 引擎自测：`test-backend-ops test -o MUL_MAT -p "rocmfpx"` 61/61、`-p "rocmfp4"`
  24/24、`-o MUL_MAT` 1382/1382、`-o MUL_MAT_ID` 935/935，全部通过，无既有失败。
- 新增回归用例（tests/test-backend-ops.cpp）：`Q8_0_ROCMFPX, m=151669, n=16384,
  k=32`（n=16384 使 17 个列 tile 溢出、NMSE 0.165 远高于 5e-4 阈值；最小形状
  n=14209 仅 1 列溢出、低于阈值故不用）。修前该用例 `ERR = 0.1649 > 5e-4` 必失败，
  修后通过。
- A 机端到端（无阀 + `-b/-ub 32768`，`/proc/<pid>/environ` 证实无
  `GGML_HIP_NO_ROCMFPX_MMQ`）：阶梯 12K/16K/20K/24K 单文档全部 200；多文档大批量
  （~24K）连续 4 次全部 200；无任何崩溃。

上游同型：`official/master`（9a9f939b9）在 `mmq.cuh:1005/1093/1182/1325` 存在同样
的 `int offset_dst = … jt*J*stride_col_dst` int32 溢出；fork 经 8cbe3f39a 为
ROCmFPX 接通 MMQ 后才将其暴露。

本提交仅本地创建，不 push。

+++ ggml/src/ggml-cuda/mmq.cuh
@@ -1091,7 +1091,7 @@ static __global__ void mul_mat_q(
         int col_high   = ncols_dst;
         int col_diff   = ncols_dst;
         int offset_y       = wt*stride_sample_y   + zt*stride_channel_y;
-        int offset_dst     = wt*stride_sample_dst + zt*stride_channel_dst + jt*J*stride_col_dst;
+        int64_t offset_dst = wt*stride_sample_dst + zt*stride_channel_dst + (int64_t) jt*J*stride_col_dst;
         int offset_y_scale;
         if constexpr (type == GGML_TYPE_NVFP4) {
             offset_y_scale = wt*nchannels_y.z*ncols_y + zt*ncols_y;
@@ -1179,7 +1179,7 @@ static __global__ void mul_mat_q(
         int col_high   = ncols_dst;
         int col_diff   = ncols_dst;
         int offset_y       = wt*stride_sample_y   + zt*stride_channel_y;
-        int offset_dst     = wt*stride_sample_dst + zt*stride_channel_dst + jt*J*stride_col_dst;
+        int64_t offset_dst = wt*stride_sample_dst + zt*stride_channel_dst + (int64_t) jt*J*stride_col_dst;
         int offset_y_scale;
         if constexpr (type == GGML_TYPE_NVFP4) {
             offset_y_scale = wt*nchannels_y.z*ncols_y + zt*ncols_y;
@@ -1268,7 +1268,7 @@ static __global__ void mul_mat_q(
     int col_high   = ncols_dst;
     int col_diff   = ncols_dst;
     int offset_y       = wt*stride_sample_y   + zt*stride_channel_y;
-    int offset_dst     = wt*stride_sample_dst + zt*stride_channel_dst + jt*J*stride_col_dst;
+    int64_t offset_dst = wt*stride_sample_dst + zt*stride_channel_dst + (int64_t) jt*J*stride_col_dst;
     int offset_y_scale;
     if constexpr (type == GGML_TYPE_NVFP4) {
         offset_y_scale = wt*nchannels_y.z*ncols_y + zt*ncols_y;
@@ -1411,7 +1411,7 @@ static __global__ void mul_mat_q_stream_k_fixup(
     const int it = tmp2.x;
 
     if (!ids_dst) {
-        const int offset_dst = wt*stride_sample_dst + zt*stride_channel_dst + jt*J*stride_col_dst + it*I;
+        const int64_t offset_dst = wt*stride_sample_dst + zt*stride_channel_dst + (int64_t) jt*J*stride_col_dst + it*I;
         dst += offset_dst;
 
         const int i_max = nrows_x   - it*I - 1;
